requirements-ml.txt
-r requirements.txt
scikit-learn>=1.5.0
joblib>=1.4.0

# Optional speech-to-text
openai-whisper>=20240930
imageio-ffmpeg>=0.5.1
transformers>=4.44.0
datasets>=2.20.0
accelerate>=0.33.0
evaluate>=0.4.2
jiwer>=3.0.4
soundfile>=0.12.1
librosa>=0.10.2
